allow only specific resolutions for custom icons

+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/branding/CustomBrandingPatch.kt
@@ -1,12 +1,13 @@
 package app.revanced.patches.youtube.layout.branding
 
 import app.revanced.patches.shared.layout.branding.baseCustomBrandingPatch
+import app.revanced.patches.shared.layout.branding.mipmapDirectories
 import java.nio.file.Files
 
 private const val APP_NAME = "YouTube ReVanced"
 
-private const val ADAPTIVE_BACKGROUND_RESOURCE_FILE_NAME = "adaptiveproduct_youtube_background_color_108.xml"
-private const val ADAPTIVE_FOREGROUND_RESOURCE_FILE_NAME = "adaptiveproduct_youtube_foreground_color_108.xml"
+private const val ADAPTIVE_BACKGROUND_RESOURCE_NAME = "adaptiveproduct_youtube_background_color_108"
+private const val ADAPTIVE_FOREGROUND_RESOURCE_NAME = "adaptiveproduct_youtube_foreground_color_108"
 
 @Suppress("unused")
 val customBrandingPatch = baseCustomBrandingPatch(
@@ -19,22 +20,20 @@ val customBrandingPatch = baseCustomBrandingPatch(
     ),
     resourceFolder = "custom-branding/youtube",
     iconResourceFileNames = arrayOf(
-        "adaptiveproduct_youtube_background_color_108",
-        "adaptiveproduct_youtube_foreground_color_108",
-        "ic_launcher",
-        "ic_launcher_round",
+        "$ADAPTIVE_BACKGROUND_RESOURCE_NAME.png",
+        "$ADAPTIVE_FOREGROUND_RESOURCE_NAME.png",
     ),
     monochromeIconFileNames = arrayOf(
         "adaptive_monochrome_ic_youtube_launcher.xml",
         "ringo2_adaptive_monochrome_ic_youtube_launcher.xml"
     ),
     adaptiveIconFileNames = arrayOf(
-        ADAPTIVE_BACKGROUND_RESOURCE_FILE_NAME,
-        ADAPTIVE_FOREGROUND_RESOURCE_FILE_NAME,
+        "$ADAPTIVE_BACKGROUND_RESOURCE_NAME.xml",
+        "$ADAPTIVE_FOREGROUND_RESOURCE_NAME.xml",
     ),
     legacyIconResourceFileNames = arrayOf(
-        "ic_launcher",
-        // "ic_launcher_round" also exists in 19.34, but was removed in later targets.
+        "ic_launcher.png",
+        // "ic_launcher_round" exists in 19.34, but was removed in later targets.
     ),
 
     block = {
@@ -49,17 +48,35 @@ val customBrandingPatch = baseCustomBrandingPatch(
     },
 
     executeBlock = {
-        val resourceDirectory = get("res/mipmap-anydpi")
+        val resourceDirectory = get("res")
 
         // Copy adaptive icon to secondary adaptive file.
         arrayOf(
-            ADAPTIVE_BACKGROUND_RESOURCE_FILE_NAME to "adaptiveproduct_youtube_2024_q4_background_color_108.xml",
-            ADAPTIVE_FOREGROUND_RESOURCE_FILE_NAME to "adaptiveproduct_youtube_2024_q4_foreground_color_108.xml",
+            ADAPTIVE_BACKGROUND_RESOURCE_NAME to "adaptiveproduct_youtube_2024_q4_background_color_108",
+            ADAPTIVE_FOREGROUND_RESOURCE_NAME to "adaptiveproduct_youtube_2024_q4_foreground_color_108",
         ).forEach { (old, new) ->
-            val oldFile = resourceDirectory.resolve(old)
-            val newFile = resourceDirectory.resolve(new)
+            var resourceType = "mipmap-anydpi"
+            val oldFile = resourceDirectory.resolve("$resourceType/$old.xml")
+            if (oldFile.exists()) {
+                val newFile = resourceDirectory.resolve("$resourceType/$new.xml")
+                Files.write(newFile.toPath(), oldFile.readBytes())
+            }
+        }
 
-            Files.write(newFile.toPath(), oldFile.readBytes())
+        // Copy mipmaps to secondary files.
+        mipmapDirectories.forEach { directory ->
+            val targetDirectory = resourceDirectory.resolve(directory)
+
+            arrayOf(
+                ADAPTIVE_BACKGROUND_RESOURCE_NAME to "adaptiveproduct_youtube_2024_q4_background_color_108",
+                ADAPTIVE_FOREGROUND_RESOURCE_NAME to "adaptiveproduct_youtube_2024_q4_foreground_color_108",
+            ).forEach { (old, new) ->
+                val oldFile = targetDirectory.resolve("$old.png")
+                if (oldFile.exists()) {
+                    val newFile = targetDirectory.resolve("$new.png")
+                    Files.write(newFile.toPath(), oldFile.readBytes())
+                }
+            }
         }
     }
 )
